+++ libs/utils/src/Response.ts
@@ -1,0 +1,9 @@
+import type { HttpResponse } from './HttpResponse.ts'
+
+/**
+ * Response API.
+ *
+ * @public
+ * @deprecated Use {@link HttpResponse} instead.
+ */
+export type Response = HttpResponse