+++ libs/utils/src/ResponseTypeMap.ts
@@ -9,7 +9,7 @@ export type ResponseTypeMap<T extends string | undefined> =
 	T extends 'json' ? any :
 	T extends 'text' ? string :
 	T extends 'blob' ? Blob :
-	T extends 'arraybuffer' ? ArrayBuffer :
+	T extends 'arrayBuffer' ? ArrayBuffer :
 	T extends 'document' ? XmlNode :
 	T extends 'stream' ? ReadableStream :
 	unknown;